refactor: move query to new unscoped transaction repository
