+++ services/121-service/src/payments/transactions/transactions.module.ts
@@ -9,7 +9,8 @@ import { TwilioMessageEntity } from '@121-service/src/notifications/twilio.entit
 import { LatestTransactionEntity } from '@121-service/src/payments/transactions/latest-transaction.entity';
 import { LatestTransactionRepository } from '@121-service/src/payments/transactions/repositories/latest-transaction.repository';
 import { TransactionEntity } from '@121-service/src/payments/transactions/transaction.entity';
-import { TransactionScopedRepository } from '@121-service/src/payments/transactions/transaction.repository';
+import { TransactionRepository } from '@121-service/src/payments/transactions/transaction.repository';
+import { TransactionScopedRepository } from '@121-service/src/payments/transactions/transaction.scoped.repository';
 import { TransactionsService } from '@121-service/src/payments/transactions/transactions.service';
 import { ProgramEntity } from '@121-service/src/programs/program.entity';
 import { RegistrationUtilsModule } from '@121-service/src/registration/modules/registration-utilts/registration-utils.module';
@@ -38,6 +39,7 @@ import { createScopedRepositoryProvider } from '@121-service/src/utils/scope/cre
   providers: [
     TransactionsService,
     TransactionScopedRepository,
+    TransactionRepository,
     RegistrationScopedRepository,
     createScopedRepositoryProvider(TransactionEntity),
     LatestTransactionRepository,
@@ -45,6 +47,7 @@ import { createScopedRepositoryProvider } from '@121-service/src/utils/scope/cre
   exports: [
     TransactionsService,
     TransactionScopedRepository,
+    TransactionRepository,
     LatestTransactionRepository,
   ],
 })
